Add files via upload
This is a one-page brochure to promote some popular sneaker shops in California. This is my first course project in the Introduction to Cartographic Design course. In this project, I want to understand the map projection knowledge and the cartographic generalization process. This map contains the major map that shows California, annotation of the sneaker shops, inset map that shows the location of California in the USA, etc.