Add files via upload
This is a reference map for Africa. This is the second project of the Introductory to Cartographic Design course. This map serves as a spatial catalog for readers, so the map needs to be an information-dense, one-stop graphic. I demonstrate my knowledge of typography (both appearance and placement) as well as continue to practice the map generalization process and visual hierarchy knowledge. This map contains capital cities' names of African countries, major cities' names of African countries, dessert names, river names, lakes names, bay names, ocean names, etc.